refactor move ModComponentFormStateWizard and NonScriptablePage
